agent/requirements.txt
fastapi==0.85.1
uvicorn==0.18.3
requests==2.28.2
python-socketio==5.8.0
selenium==4.5.0
pymongo>=3.11.0
pydantic==1.10.19